Updating Users, Profile Pictures, Bug fixes
Fixed/implemented user editing. Updated user gets saved so when you return to the profile page the updates persist.

Fixed the profile upload. Note: Does not work due to URL but should work in prod. Made it so, you have to click on the image to get the upload pop up instead of just hover

Made it so on login page when you press enter it logs you in instead of going to register

Cosmetic changes. Removed unnecessary comments